feat(login): sign up form UI
