+++ feature/authentication/src/main/res/values/strings.xml
@@ -3,6 +3,7 @@
     <string name="email_field_placeholder">Email</string>
     <string name="email_field_error_invalid">Email is invalid</string>
     <string name="password_field_placeholder">Password</string>
+    <string name="confirm_password_field_placeholder">Confirm Password</string>
     <string name="password_field_error_invalid">Password length must be from 4 to 20 characters and must contains lowercase, uppercase, number and special character</string>
     <string name="sign_in">Sign In</string>
     <string name="sign_up">Sign Up</string>
